+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 * Add new method :meth:`pymaf.MafFrame.calculate_concordance`.
 * Add new method :meth:`pymaf.MafFrame.copy`.
 * Add new method :meth:`pymaf.MafFrame.filter_indel`.
+* Add new method :meth:`pymaf.MafFrame.plot_comparison`.
 
 0.24.0 (2021-10-02)
 -------------------